Finalizados testes dos load/store

+++ fortnight04/cpu/basiccpu/test/BasicCPUTest.cpp
@@ -62,7 +62,7 @@ void BasicCPUTest::setW(int n, int value) {
 	BasicCPU::setW(n,value);
 }
 
-void BasicCPUTest::setX(int n, int value) {
+void BasicCPUTest::setX(int n, long value) {
 	BasicCPU::setX(n,value);
 }
